New Front page
Changed the index page of the project in order to give it a new look n
feel. older page was made with the concepts od W3Schools as at that time
just started leaning jsp n bootstrap. Now after 4 weeks learned new
tecniques and concepts so try to implement them with the new front/index
page. in further days going to change rest of the project acc. to the
new index/front page inorder to make one look n design for whole project